backend/requirements.txt
# Flask y dependencias principales
Flask==3.1.0
Flask-Cors==5.0.0
gunicorn==23.0.0

# TensorFlow y dependencias relacionadas
tensorflow==2.16.0
tensorboard==2.16.0
tensorflow-io-gcs-filesystem==0.31.0
h5py==3.12.1
keras==2.16.0

# NumPy y SciPy
numpy==1.23.5  # Versión compatible con TensorFlow, SciPy, y otras bibliotecas
scipy==1.9.3

# Scikit-learn y dependencias relacionadas
scikit-learn==1.3.2 
joblib==1.4.2
threadpoolctl==3.5.0

# Matplotlib y dependencias relacionadas
matplotlib==3.5.3
contourpy==1.0.7 
cycler==0.12.1
fonttools==4.55.3
kiwisolver==1.3.2
packaging==24.2
pillow==11.1.0
pyparsing==3.2.1

# Transformers y dependencias relacionadas
transformers==4.49.0
tokenizers==0.21.0
ml-dtypes==0.4.1
torch==2.6.0
safetensors==0.5.2

# Supabase y dependencias relacionadas
supabase==2.11.0
postgrest==0.19.1
python-dotenv==1.0.1

# Otras dependencias
requests==2.32.3
urllib3==2.3.0
aiohttp==3.11.11
async-timeout==5.0.1
attrs==24.3.0
charset-normalizer==3.4.1
colorama==0.4.6
deprecation==2.1.0
exceptiongroup==1.2.2
filelock==3.17.0
flatbuffers==24.12.23
frozenlist==1.5.0
fsspec==2025.2.0
gast==0.4.0  # Versión compatible con TensorFlow
google-pasta==0.2.0
grpcio==1.69.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.28.1
hyperframe==6.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
networkx==2.8.8 
opt-einsum==3.4.0
optree==0.14.0
protobuf==3.20.3
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
regex==2024.11.6
rich==13.9.4
sniffio==1.3.1
storage3==0.11.0
StrEnum==0.4.15
supafunc==0.9.2
sympy==1.13.1
termcolor==2.5.0
tqdm==4.67.1
typing_extensions==4.12.2
websockets==13.1
Werkzeug==3.1.3
wrapt==1.14.1
yarl==1.18.3